Connect `FileContentCompleter` to `prompt()` call in `get_input()` for autocompletion.`
# Context:
USER: connect the filecontentcompleter to the prompt() call in get_input, so that it autocompletes
ASSISTANT: To connect the `FileContentCompleter` to the `prompt()` call in the `get_input` method, you need to pass the `FileContentCompleter` instance as the `completer` argument in the `prompt()` function. Here's the change:

```python
coder.py
<<<<<<< ORIGINAL
            if multiline_input:
                line = prompt(". ")
            else:
                line = prompt("> ")
=======
            completer_instance = FileContentCompleter(self.fnames)
            if multiline_input:
                line = prompt(". ", completer=completer_instance)
            else:
                line = prompt("> ", completer=completer_instance)
>>>>>>> UPDATED
```

This change will enable autocompletion using the `FileContentCompleter` when you call `prompt()`.