aider: Changed `self.words` initialization from an empty list to an empty set and replaced the `extend` method with the `update` method to add words to the set.
# Context:
USER: self.words should be a set not a list
ASSISTANT: To change `self.words` from a list to a set, we will update the `__init__` method in the `FileContentCompleter` class. Here's the updated code:

```python
aider/getinput.py
<<<<<<< ORIGINAL
class FileContentCompleter(Completer):
    def __init__(self, fnames):
        self.words = []
        for fname in fnames:
            with open(fname, "r") as f:
                content = f.read()
            self.words.extend(re.split(r'\W+', content))
=======
class FileContentCompleter(Completer):
    def __init__(self, fnames):
        self.words = set()
        for fname in fnames:
            with open(fname, "r") as f:
                content = f.read()
            self.words.update(re.split(r'\W+', content))
>>>>>>> UPDATED
```

This change updates the `self.words` initialization from an empty list to an empty set and replaces the `extend` method with the `update` method to add words to the set.